move test related files to /test directory
